third_party/blink/web_tests/http/tests/security/sandboxed-iframe-invalid-expected.txt

CONSOLE ERROR: Error while parsing the 'sandbox' attribute: 'allowScripts' is an invalid sandbox flag.
CONSOLE ERROR: Error while parsing the 'sandbox' attribute: 'allowScripts', 'allowSameOrigin', 'allowFoobarbloop' are invalid sandbox flags.
Test that an iframe with invalid sandbox flags generates a relevant warning.

 
